update everything to beta 2.3.1
i've finished a hardcoding necessary api
and added a debugmenu thing for the first time. thank god. anyway. it
also has some of the things renamed and some new classes. big changes
in the referencing scheme too, the majority of needed ones are now
"protected" for the api to access.