removed addons system and started to find things unnecessary
the zip is the code this is based on (the mac version).
i removed the windows blocking system. i need someone to change the
paths and the required libraries for this. most things should still
work, but the majority of things are broken. i fixed InfoMethods.info
though. the differences between these are staggering.