Beginning a process to add some listeners using a different system.
still doing some minor changes to the older code base as just a fix for
things that no longer are necessary. i havent really been focussing
much on this project. if anyone wants to take it up, i'll look thru
your code and vet it a bit and commit the algorithms if i like them.