3.0_dev07
this is probably the first working version from the rewrite. its a
massive change. tons of stuff have been worked on. i added a system to
do the logging in the console with that bash prompt thing. there was
only a marginal amount of more documentation. integrated some stuff
from the sysetm read out too. solved some NPE's in the for loop for the
help list.