3.1_03
fixed some ancient code which could cause NPEs in the future.
deprecated the 'append' method, as it is no longer necessary, redid the
entirety of the getOutText method to return the TextArea instead of a
string (which allows some more interesting things), fixed another error
in the CD implementation, and changed all code to reflect that. This is
a nice set of behind the scenes bug fixes. Furthermore, we now tell you
the name of the directory you are in when you execute a command. I also
changed the autoformatter, which has necessitated a change in
practically everything that has comments.