3.3_dev05: Out With The Old, In With The New
Preliminary ClassLoading system set up (took forever). Rewrote the
entire properties system to eschew the obnoxiousness of a single text
file and go directly to a standard properties file (man, that took
forever). Also, created a plethora of new methods to handle those
properties systems, rewrote how the application handles quitting, the
CLI configuration handler, the GUI Configuration Reader, added
documentation for the API Definitions, and started writing the API
Implementation (man, will writing that implementation take a long
time…). If any bugs have been fixed, I’d be really surprised.

Note: This is effectively an ALPHA build. It is honestly terrible when
it comes to dealing with things. I’d love it if someone were to help
restructure the logic, and secondarily, it probably created a myriad of
new bugs.