[BROKEN] Started process to remove as many 'static' keywords as possible
I want to turn this bloody thing into a dynamic system so I can create
as many bloody windows as I want. That requires a near total
restructuring. This would be 3.8 if finished. API is on hold until this
is done, since to make the API work, it would require this.